+++ satella/coding/concurrent/futures/__init__.py
@@ -1,0 +1,4 @@
+from .futures import Future, WrappingFuture, InvalidStateError
+from .call_in_future import call_in_future
+
+__all__ = ['Future', 'WrappingFuture', 'InvalidStateError', 'call_in_future']