+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.12.6
 
+* added `call_in_future`
+