+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 * `call_in_future` callable will now return the Future
 * added `ExecutorWrapper`
 * added `Future.on_success` and `Future.on_failure`
-
+* `apply_dict_object` won't DictObject if passed a DictObject already